v0.33 - fixeado air1

+++ AirplaneStats/src/air1/AirTimeCombiner.java
@@ -1,16 +1,16 @@
 package air1;
 import java.io.IOException;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
-public class AirTimeCombiner extends Reducer<Text, IntWritable, Text, IntWritable> {
-	private IntWritable result = new IntWritable();
-	public void reduce(Text key, Iterable<IntWritable> values, Context context) throws
+public class AirTimeCombiner extends Reducer<Text, FloatWritable, Text, FloatWritable> {
+	private FloatWritable result = new FloatWritable();
+	public void reduce(Text key, Iterable<FloatWritable> values, Context context) throws
 	IOException, InterruptedException {
 		// combiner for each class is almost same only output value may be change as per requirement
 		int sum = 0;
-		for (IntWritable val : values) {
+		for (FloatWritable val : values) {
 			sum += val.get();
 		}
 		result.set(sum);
